+++ core/src/main/java/grammar/Grammar.java
@@ -1,4 +1,4 @@
-package lexer;
+package grammar;
 
 import automata.FiniteStateMachine;
