+++ core/src/main/java/lexer/Lexer.java
@@ -1,5 +1,7 @@
 package lexer;
 
+import grammar.LexerGrammar;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
