+++ core/src/main/java/lexer/Token.java
@@ -1,5 +1,7 @@
 package lexer;
 
+import grammar.LexerGrammar;
+
 public class Token {
 
     private LexerGrammar type;
